+++ apps/builder/src/components/ImageUploadContent/UnsplashPicker.tsx
@@ -11,6 +11,7 @@ import {
   Spinner,
   Stack,
   Text,
+  useColorModeValue,
 } from '@chakra-ui/react'
 import { env, isDefined, isEmpty } from '@typebot.io/lib'
 import { useCallback, useEffect, useRef, useState } from 'react'
@@ -30,6 +31,7 @@ type Props = {
 }
 
 export const UnsplashPicker = ({ imageSize, onImageSelect }: Props) => {
+  const unsplashLogoFillColor = useColorModeValue('black', 'white')
   const [isFetching, setIsFetching] = useState(false)
   const [images, setImages] = useState<UnsplashImage[]>([])
   const [error, setError] = useState<string | null>(null)
@@ -40,7 +42,6 @@ export const UnsplashPicker = ({ imageSize, onImageSelect }: Props) => {
   const [nextPage, setNextPage] = useState(0)
 
   const fetchNewImages = useCallback(async (query: string, page: number) => {
-    console.log('Fetch images', query, page)
     if (query === '') return searchRandomImages()
     if (query.length <= 2) return
     setError(null)
@@ -74,7 +75,6 @@ export const UnsplashPicker = ({ imageSize, onImageSelect }: Props) => {
     if (!bottomAnchor.current) return
     const observer = new IntersectionObserver(
       (entities: IntersectionObserverEntry[]) => {
-        console.log('Intersection observer', entities)
         const target = entities[0]
         if (target.isIntersecting) fetchNewImages(searchQuery, nextPage + 1)
       },
@@ -146,7 +146,7 @@ export const UnsplashPicker = ({ imageSize, onImageSelect }: Props) => {
             'UNSPLASH_APP_NAME'
           )}&utm_medium=referral`}
         >
-          <UnsplashLogo width="80px" />
+          <UnsplashLogo width="80px" fill={unsplashLogoFillColor} />
         </Link>
       </HStack>
       {isDefined(error) && (
@@ -190,6 +190,7 @@ type UnsplashImageProps = {
 }
 
 const UnsplashImage = ({ image, onClick }: UnsplashImageProps) => {
+  const linkColor = useColorModeValue('gray.500', 'gray.400')
   const { user, urls, alt_description } = image
 
   return (
@@ -210,7 +211,7 @@ const UnsplashImage = ({ image, onClick }: UnsplashImageProps) => {
           'UNSPLASH_APP_NAME'
         )}&utm_medium=referral`}
         noOfLines={1}
-        color="gray.500"
+        color={linkColor}
       >
         {user.name}
       </TextLink>