+++ apps/builder/src/features/editor/components/BlockIcon.tsx
@@ -37,6 +37,7 @@ import { ConditionIcon } from '@/features/blocks/logic/condition/components/Cond
 import { RedirectIcon } from '@/features/blocks/logic/redirect/components/RedirectIcon'
 import { SetVariableIcon } from '@/features/blocks/logic/setVariable/components/SetVariableIcon'
 import { TypebotLinkIcon } from '@/features/blocks/logic/typebotLink/components/TypebotLinkIcon'
+import { AbTestIcon } from '@/features/blocks/logic/abTest/components/AbTestIcon'
 
 type BlockIconProps = { type: BlockType } & IconProps
 
@@ -44,6 +45,8 @@ export const BlockIcon = ({ type, ...props }: BlockIconProps): JSX.Element => {
   const blue = useColorModeValue('blue.500', 'blue.300')
   const orange = useColorModeValue('orange.500', 'orange.300')
   const purple = useColorModeValue('purple.500', 'purple.300')
+  const openAIColor = useColorModeValue('black', 'white')
+
   switch (type) {
     case BubbleBlockType.TEXT:
       return <TextBubbleIcon color={blue} {...props} />
@@ -89,6 +92,8 @@ export const BlockIcon = ({ type, ...props }: BlockIconProps): JSX.Element => {
       return <JumpIcon color={purple} {...props} />
     case LogicBlockType.TYPEBOT_LINK:
       return <TypebotLinkIcon color={purple} {...props} />
+    case LogicBlockType.AB_TEST:
+      return <AbTestIcon color={purple} {...props} />
     case IntegrationBlockType.GOOGLE_SHEETS:
       return <GoogleSheetsLogo {...props} />
     case IntegrationBlockType.GOOGLE_ANALYTICS:
@@ -106,7 +111,7 @@ export const BlockIcon = ({ type, ...props }: BlockIconProps): JSX.Element => {
     case IntegrationBlockType.CHATWOOT:
       return <ChatwootLogo {...props} />
     case IntegrationBlockType.OPEN_AI:
-      return <OpenAILogo {...props} />
+      return <OpenAILogo fill={openAIColor} {...props} />
     case 'start':
       return <FlagIcon {...props} />
   }