+++ apps/builder/src/features/results/components/ResultsTableContainer.tsx
@@ -1,12 +1,14 @@
 import { Stack } from '@chakra-ui/react'
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { LogsModal } from './LogsModal'
 import { useTypebot } from '@/features/editor/providers/TypebotProvider'
 import { useResults } from '../ResultsProvider'
 import { ResultModal } from './ResultModal'
 import { ResultsTable } from './table/ResultsTable'
+import { useRouter } from 'next/router'
 
 export const ResultsTableContainer = () => {
+  const { query } = useRouter()
   const {
     flatResults: results,
     fetchNextPage,
@@ -18,21 +20,25 @@ export const ResultsTableContainer = () => {
   const [inspectingLogsResultId, setInspectingLogsResultId] = useState<
     string | null
   >(null)
-  const [expandedResultIndex, setExpandedResultIndex] = useState<number | null>(
-    null
-  )
+  const [expandedResultId, setExpandedResultId] = useState<string | null>(null)
 
   const handleLogsModalClose = () => setInspectingLogsResultId(null)
 
-  const handleResultModalClose = () => setExpandedResultIndex(null)
+  const handleResultModalClose = () => setExpandedResultId(null)
 
   const handleLogOpenIndex = (index: number) => () => {
     if (!results[index]) return
     setInspectingLogsResultId(results[index].id)
   }
 
-  const handleResultExpandIndex = (index: number) => () =>
-    setExpandedResultIndex(index)
+  const handleResultExpandIndex = (index: number) => () => {
+    if (!results[index]) return
+    setExpandedResultId(results[index].id)
+  }
+
+  useEffect(() => {
+    if (query.id) setExpandedResultId(query.id as string)
+  }, [query.id])
 
   return (
     <Stack pb="28" px={['4', '0']} spacing="4" maxW="1600px" w="full">
@@ -44,7 +50,7 @@ export const ResultsTableContainer = () => {
         />
       )}
       <ResultModal
-        resultIdx={expandedResultIndex}
+        resultId={expandedResultId}
         onClose={handleResultModalClose}
       />