+++ packages/schemas/features/chat.ts
@@ -73,9 +73,7 @@ const chatSessionSchema = z.object({
 
 const textMessageSchema = z.object({
   type: z.literal(BubbleBlockType.TEXT),
-  content: textBubbleContentSchema.omit({
-    richText: true,
-  }),
+  content: textBubbleContentSchema,
 })
 
 const imageMessageSchema = z.object({
@@ -220,6 +218,11 @@ const clientSideActionSchema = z
           wait: z.object({
             secondsToWaitFor: z.number(),
           }),
+        })
+      )
+      .or(
+        z.object({
+          setVariable: z.object({ scriptToExecute: scriptToExecuteSchema }),
         })
       )
   )